chore: 폴더 구조 변경, 필요없는 파일 삭제 	modified:   src/screens/HomeScreen/CurrentForecastScreen/CurrentForecastScreen.tsx 	renamed:    src/screens/HomeScreen/_components/weatherInfoHeader/WeatherInfoHeader.tsx -> src/screens/HomeScreen/CurrentForecastScreen/_components/weatherInfoHeader/WeatherInfoHeader.tsx 	renamed:    src/screens/HomeScreen/_components/weatherInfoHeader/WeatherInfoHeader.type.tsx -> src/screens/HomeScreen/CurrentForecastScreen/_components/weatherInfoHeader/WeatherInfoHeader.type.tsx 	renamed:    src/screens/HomeScreen/_components/weatherInfoHeader/WeatherInfoHeaderIcon.tsx -> src/screens/HomeScreen/CurrentForecastScreen/_components/weatherInfoHeader/WeatherInfoHeaderIcon.tsx 	renamed:    src/screens/HomeScreen/_components/weatherInfoHeader/WeatherInfoHeaderSummary.tsx -> src/screens/HomeScreen/CurrentForecastScreen/_components/weatherInfoHeader/WeatherInfoHeaderSummary.tsx 	renamed:    src/screens/HomeScreen/_components/weatherInfoHeader/WeatherInfoHeaderTempSection.tsx -> src/screens/HomeScreen/CurrentForecastScreen/_components/weatherInfoHeader/WeatherInfoHeaderTempSection.tsx 	renamed:    src/screens/HomeScreen/_components/weatherInfoHeader/LocationHeader.const.tsx -> src/screens/HomeScreen/_components/locationHeader/LocationHeader.const.tsx 	renamed:    src/screens/HomeScreen/_components/weatherInfoHeader/LocationHeader.tsx -> src/screens/HomeScreen/_components/locationHeader/LocationHeader.tsx 	renamed:    src/screens/HomeScreen/_components/weatherInfoHeader/__test__/LocationHeader.test.tsx -> src/screens/HomeScreen/_components/locationHeader/__test__/LocationHeader.test.tsx 	deleted:    src/screens/HomeScreen/_components/weatherInfoHeader/WeatherInfoHeader.const.tsx
