+++ src/screens/HomeScreen/CurrentForecastScreen/CurrentForecastScreen.tsx
@@ -5,7 +5,7 @@ import { useNavigation } from '@react-navigation/native';
 import TestSection from '@components/TestSection';
 import { HOME_NAVIGATION_TEST_ID_LIST } from '@navigation/home/HomeNavigation.const';
 import HomeScreenWrapper from '@screens/HomeScreen/_components/HomeScreenWrapper';
-import WeatherInfoHeader from '@screens/HomeScreen/_components/weatherInfoHeader/WeatherInfoHeader';
+import WeatherInfoHeader from '@screens/HomeScreen/CurrentForecastScreen/_components/weatherInfoHeader/WeatherInfoHeader';
 
 export default function CurrentForecastScreen() {
   const { navigate } = useNavigation();
