+++ src/screens/HomeScreen/CurrentForecastScreen/_components/weatherInfoHeader/WeatherInfoHeader.type.tsx
