+++ src/screens/HomeScreen/CurrentForecastScreen/_components/weatherInfoHeader/WeatherInfoHeaderIcon.tsx
