+++ src/screens/HomeScreen/CurrentForecastScreen/_components/weatherInfoHeader/WeatherInfoHeaderTempSection.tsx
