+++ src/screens/HomeScreen/_components/locationHeader/LocationHeader.const.tsx
