--- src/screens/HomeScreen/_components/weatherInfoHeader/WeatherInfoHeader.const.tsx
@@ -1,15 +1,0 @@
-export const WEATHER_HEADER_ICON_SIZE = 84;
-export const WEATHER_HEADER_ICON_SIZE_SCALE = 140;
-
-export const WEATHER_HEADER_HEIGHT = 180;
-export const WEATHER_HEADER_HEIGHT_SCALE = 342;
-
-export const CURRENT_TEMP_SIZE = 32;
-export const CURRENT_TEMP_SIZE_SCALE = 52;
-
-export const SUMMARY_SIZE = 14;
-export const SUMMARY_SIZE_SCALE = 16;
-
-export const MAX_SCROLL_VALUE = WEATHER_HEADER_HEIGHT_SCALE - WEATHER_HEADER_HEIGHT;
-
-export const WEATHER_INFO_HEADER_ANIMATION_DURATION = 50;
